MOVEONLY: Consensus: Policy: Move CFeeRate from amount.o to policy/fees.o
