+++ src/init.cpp
@@ -18,6 +18,7 @@
 #include "main.h"
 #include "miner.h"
 #include "net.h"
+#include "policy/fees.h"
 #include "policy/policy.h"
 #include "rpcserver.h"
 #include "script/standard.h"
