+++ src/qt/sendcoinsdialog.cpp
@@ -18,6 +18,7 @@
 #include "base58.h"
 #include "coincontrol.h"
 #include "main.h"
+#include "policy/fees.h"
 #include "ui_interface.h"
 #include "wallet/wallet.h"
