+++ src/rpcnet.cpp
@@ -8,6 +8,7 @@
 #include "main.h"
 #include "net.h"
 #include "netbase.h"
+#include "policy/fees.h"
 #include "protocol.h"
 #include "sync.h"
 #include "timedata.h"
